+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,22 @@
 # gt-react-native
 
+## 10.19.19
+
+### Patch Changes
+
+- Updated dependencies [[`e041312`](https://github.com/generaltranslation/gt/commit/e04131263dd61e469db977bcc196dc1283e773d0)]:
+  - generaltranslation@8.2.16
+  - @generaltranslation/react-core@1.8.21
+  - @generaltranslation/supported-locales@2.1.1
+
+## 10.19.18
+
+### Patch Changes
+
+- Updated dependencies [[`cc4edc1`](https://github.com/generaltranslation/gt/commit/cc4edc1e40d9977125bf6d053fb7f8cdbdd40b05)]:
+  - @generaltranslation/supported-locales@2.1.0
+  - @generaltranslation/react-core@1.8.20
+
 ## 10.19.17
 
 ### Patch Changes